+++ packages/core-js/modules/esnext.uint8-array.to-base64.js
@@ -4,6 +4,7 @@ var global = require('../internals/global');
 var uncurryThis = require('../internals/function-uncurry-this');
 var anObjectOrUndefined = require('../internals/an-object-or-undefined');
 var anUint8Array = require('../internals/an-uint8-array');
+var notDetached = require('../internals/array-buffer-not-detached');
 var base64Map = require('../internals/base64-map');
 var getAlphabetOption = require('../internals/get-alphabet-option');
 
@@ -21,6 +22,7 @@ if (Uint8Array) $({ target: 'Uint8Array', proto: true }, {
     var options = arguments.length ? anObjectOrUndefined(arguments[0]) : undefined;
     var alphabet = getAlphabetOption(options) === 'base64' ? base64Alphabet : base64UrlAlphabet;
     var omitPadding = !!options && !!options.omitPadding;
+    notDetached(this.buffer);
 
     var result = '';
     var i = 0;