+++ tests/unit-global/esnext.uint8-array.to-hex.js
@@ -9,6 +9,13 @@ if (DESCRIPTORS) QUnit.test('Uint8Array.prototype.toHex', assert => {
 
   assert.same(new Uint8Array([72, 101, 108, 108, 111, 32, 87, 111, 114, 108, 100]).toHex(), '48656c6c6f20576f726c64', 'proper result');
 
+  if (ArrayBuffer.prototype.transfer) {
+    const array = new Uint8Array([72, 101, 108, 108, 111, 32, 87, 111, 114, 108, 100]);
+    array.buffer.transfer();
+
+    assert.throws(() => array.toHex(), TypeError, 'detached');
+  }
+
   assert.throws(() => toHex.call(null), TypeError, "isn't generic #1");
   assert.throws(() => toHex.call(undefined), TypeError, "isn't generic #2");
   assert.throws(() => toHex.call(new Int16Array([1])), TypeError, "isn't generic #3");